Implement Control Protocol v2
- Rename v1.1 Control Protocol to v2
- Present bundle.conf found within bundle zip file as part of the conduct load request as opposed to parsing the bundle.conf for fields declared within the file.
- Present the overlay bundle.conf if found within the configuration zip in the similar fashion
- Also rename Control Protocol v1.0 to v1
