+++ conductr_cli/exceptions.py
@@ -12,3 +12,11 @@ class Boot2DockerError(Exception):
 
     def __str__(self):
         return repr(self.value)
+
+
+class MalformedBundleError(Exception):
+    def __init__(self, value):
+        self.value = value
+
+    def __str__(self):
+        return repr(self.value)
