+++ conductr_cli/test/conduct_load_test_base.py
@@ -1,14 +1,12 @@
-from conductr_cli.test.cli_test_case import CliTestCase, create_temp_bundle, create_temp_bundle_with_contents, \
-    strip_margin, as_error
+from conductr_cli.test.cli_test_case import CliTestCase, strip_margin, as_error
 from conductr_cli import conduct_load
 from conductr_cli.conduct_load import LOAD_HTTP_TIMEOUT
 from urllib.error import URLError
-import shutil
 
 try:
-    from unittest.mock import call, patch, MagicMock  # 3.3 and beyond
+    from unittest.mock import patch, MagicMock  # 3.3 and beyond
 except ImportError:
-    from mock import call, patch, MagicMock
+    from mock import patch, MagicMock
 
 
 class ConductLoadTestBase(CliTestCase):
@@ -123,37 +121,6 @@ class ConductLoadTestBase(CliTestCase):
             self.default_output(params=cli_parameters),
             self.output(stdout))
 
-    def base_test_success_with_configuration(self):
-        tmpdir, config_file = create_temp_bundle_with_contents({
-            'bundle.conf': '{name="overlaid-name"}',
-            'config.sh': 'echo configuring'
-        })
-
-        urlretrieve_mock = MagicMock(side_effect=[(self.bundle_file, ()), (config_file, ())])
-        http_method = self.respond_with(200, self.default_response)
-        stdout = MagicMock()
-        open_mock = MagicMock(return_value=1)
-
-        with patch('conductr_cli.conduct_load.urlretrieve', urlretrieve_mock), \
-                patch('requests.post', http_method), \
-                patch('sys.stdout', stdout), \
-                patch('builtins.open', open_mock):
-            args = self.default_args.copy()
-            args.update({'configuration': config_file})
-            conduct_load.load(MagicMock(**args))
-
-        self.assertEqual(
-            open_mock.call_args_list,
-            [call(self.bundle_file, 'rb'), call(config_file, 'rb')]
-        )
-
-        expected_files = self.default_files + [('configuration', ('bundle.zip', 1))]
-        expected_files[4] = ('bundleName', 'overlaid-name')
-        http_method.assert_called_with(self.default_url, files=expected_files, timeout=LOAD_HTTP_TIMEOUT)
-
-        self.assertEqual(self.default_output(downloading_configuration='Retrieving configuration...\n'),
-                         self.output(stdout))
-
     def base_test_failure(self):
         urlretrieve_mock = MagicMock(return_value=(self.bundle_file, ()))
         http_method = self.respond_with(404)
@@ -192,117 +159,6 @@ class ConductLoadTestBase(CliTestCase):
         self.assertEqual(
             self.default_connection_error.format(self.default_url),
             self.output(stderr))
-
-    def base_test_failure_no_nr_of_cpus(self):
-        stderr = MagicMock()
-
-        tmpdir, bundle_file = create_temp_bundle(
-            strip_margin("""|memory     = {}
-                            |diskSpace  = {}
-                            |roles      = [{}]
-                            |""").format(self.memory, self.disk_space, ', '.join(self.roles)))
-
-        with patch('sys.stderr', stderr):
-            args = self.default_args.copy()
-            args.update({'bundle': bundle_file})
-            conduct_load.load(MagicMock(**args))
-
-        self.assertEqual(
-            as_error(strip_margin("""|Error: Unable to parse bundle.conf.
-                                     |Error: No configuration setting found for key nrOfCpus.
-                                     |""")),
-            self.output(stderr))
-
-        shutil.rmtree(tmpdir)
-
-    def base_test_failure_no_memory(self):
-        stderr = MagicMock()
-
-        tmpdir, bundle_file = create_temp_bundle(
-            strip_margin("""|nrOfCpus   = {}
-                            |diskSpace  = {}
-                            |roles      = [{}]
-                            |""").format(self.nr_of_cpus, self.disk_space, ', '.join(self.roles)))
-
-        with patch('sys.stderr', stderr):
-            args = self.default_args.copy()
-            args.update({'bundle': bundle_file})
-            conduct_load.load(MagicMock(**args))
-
-        self.assertEqual(
-            as_error(strip_margin("""|Error: Unable to parse bundle.conf.
-                                     |Error: No configuration setting found for key memory.
-                                     |""")),
-            self.output(stderr))
-
-        shutil.rmtree(tmpdir)
-
-    def base_test_failure_no_disk_space(self):
-        stderr = MagicMock()
-
-        tmpdir, bundle_file = create_temp_bundle(
-            strip_margin("""|nrOfCpus   = {}
-                            |memory     = {}
-                            |roles      = [{}]
-                            |""").format(self.nr_of_cpus, self.memory, ', '.join(self.roles)))
-
-        with patch('sys.stderr', stderr):
-            args = self.default_args.copy()
-            args.update({'bundle': bundle_file})
-            conduct_load.load(MagicMock(**args))
-
-        self.assertEqual(
-            as_error(strip_margin("""|Error: Unable to parse bundle.conf.
-                                     |Error: No configuration setting found for key diskSpace.
-                                     |""")),
-            self.output(stderr))
-
-        shutil.rmtree(tmpdir)
-
-    def base_test_failure_no_roles(self):
-        stderr = MagicMock()
-
-        tmpdir, bundle_file = create_temp_bundle(
-            strip_margin("""|nrOfCpus   = {}
-                            |memory     = {}
-                            |diskSpace  = {}
-                            |""").format(self.nr_of_cpus, self.memory, self.disk_space))
-
-        with patch('sys.stderr', stderr):
-            args = self.default_args.copy()
-            args.update({'bundle': bundle_file})
-            conduct_load.load(MagicMock(**args))
-
-        self.assertEqual(
-            as_error(strip_margin("""|Error: Unable to parse bundle.conf.
-                                     |Error: No configuration setting found for key roles.
-                                     |""")),
-            self.output(stderr))
-
-        shutil.rmtree(tmpdir)
-
-    def base_test_failure_roles_not_a_list(self):
-        stderr = MagicMock()
-
-        tmpdir, bundle_file = create_temp_bundle(
-            strip_margin("""|nrOfCpus   = {}
-                            |memory     = {}
-                            |diskSpace  = {}
-                            |roles      = {}
-                            |""").format(self.nr_of_cpus, self.memory, self.disk_space, '-'.join(self.roles)))
-
-        with patch('sys.stderr', stderr):
-            args = self.default_args.copy()
-            args.update({'bundle': bundle_file})
-            conduct_load.load(MagicMock(**args))
-
-        self.assertEqual(
-            as_error(strip_margin("""|Error: Unable to parse bundle.conf.
-                                     |Error: roles has type 'str' rather than 'list'.
-                                     |""")),
-            self.output(stderr))
-
-        shutil.rmtree(tmpdir)
 
     def base_test_failure_no_bundle(self):
         urlretrieve_mock = MagicMock(side_effect=URLError('no_such.bundle'))
