+++ conductr_cli/test/test_conduct_logs.py
@@ -14,7 +14,7 @@ class TestConductLogsCommand(CliTestCase):
     default_args = {
         'ip': '127.0.0.1',
         'port': '9005',
-        'api_version': '1.0',
+        'api_version': '1',
         'bundle': 'ab8f513',
         'lines': 1,
         'date': True,
