+++ conductr_cli/test/test_conduct_run_v1.py
@@ -1,5 +1,7 @@
+from conductr_cli.test.cli_test_case import strip_margin
 from conductr_cli.test.conduct_run_test_base import ConductRunTestBase
 from conductr_cli import conduct_run
+
 
 try:
     from unittest.mock import patch, MagicMock  # 3.3 and beyond
@@ -15,7 +17,7 @@ class TestConductRunCommand(ConductRunTestBase):
         self.default_args = {
             'ip': '127.0.0.1',
             'port': 9005,
-            'api_version': '1.1',
+            'api_version': '1',
             'verbose': False,
             'long_ids': False,
             'cli_parameters': '',
@@ -24,7 +26,7 @@ class TestConductRunCommand(ConductRunTestBase):
             'affinity': None
         }
 
-        self.default_url = 'http://127.0.0.1:9005/v1.1/bundles/45e0c477d3e5ea92aa8d85c0d8f3e25c?scale=3'
+        self.default_url = 'http://127.0.0.1:9005/bundles/45e0c477d3e5ea92aa8d85c0d8f3e25c?scale=3'
 
         self.output_template = """|Bundle run request sent.
                                   |Stop bundle with: conduct stop{params} {bundle_id}
@@ -49,11 +51,11 @@ class TestConductRunCommand(ConductRunTestBase):
     def test_failure_invalid_address(self):
         self.base_test_failure_invalid_address()
 
-    def test_success_with_affinity(self):
+    def test_error_with_affinity_switch(self):
         args = {
             'ip': '127.0.0.1',
             'port': 9005,
-            'api_version': '1.1',
+            'api_version': '1.0',
             'verbose': False,
             'long_ids': False,
             'cli_parameters': '',
@@ -62,15 +64,11 @@ class TestConductRunCommand(ConductRunTestBase):
             'affinity': 'other-bundle'
         }
 
-        expected_url = \
-            'http://127.0.0.1:9005/v1.1/bundles/45e0c477d3e5ea92aa8d85c0d8f3e25c?scale=3&affinity=other-bundle'
-
-        http_method = self.respond_with(200, self.default_response)
-        stdout = MagicMock()
-
-        with patch('requests.put', http_method), patch('sys.stdout', stdout):
+        stderr = MagicMock()
+        with patch('sys.stderr', stderr):
             conduct_run.run(MagicMock(**args))
 
-        http_method.assert_called_with(expected_url)
-
-        self.assertEqual(self.default_output(), self.output(stdout))
+        self.assertEqual(
+            strip_margin("""|ERROR: Affinity feature is only available for v1.1 onwards of ConductR
+                            |"""),
+            self.output(stderr))
