+++ conductr_cli/test/test_conduct_stop.py
@@ -21,7 +21,7 @@ class TestConductStopCommand(CliTestCase):
     default_args = {
         'ip': '127.0.0.1',
         'port': 9005,
-        'api_version': '1.0',
+        'api_version': '1',
         'verbose': False,
         'long_ids': False,
         'cli_parameters': '',
