+++ conductr_cli/test/test_conduct_url.py
@@ -9,18 +9,18 @@ except ImportError:
 
 class TestConductUrl(TestCase):
 
-    def test_url_v1_0(self):
+    def test_url_v1(self):
         args = MagicMock()
         args.ip = '127.0.0.1'
         args.port = 9005
-        args.api_version = '1.0'
+        args.api_version = '1'
         result = conduct_url.url('test', args)
         self.assertEqual('http://127.0.0.1:9005/test', result)
 
-    def test_url_v1_1(self):
+    def test_url_v2(self):
         args = MagicMock()
         args.ip = '127.0.0.1'
         args.port = 9005
-        args.api_version = '1.1'
+        args.api_version = '2'
         result = conduct_url.url('test', args)
-        self.assertEqual('http://127.0.0.1:9005/v1.1/test', result)
+        self.assertEqual('http://127.0.0.1:9005/v2/test', result)
