+++ conductr_cli/validation.py
@@ -11,7 +11,7 @@ from urllib.error import URLError
 from zipfile import BadZipFile
 from conductr_cli import terminal
 from conductr_cli.ansi_colors import RED, YELLOW, UNDERLINE, ENDC
-from conductr_cli.exceptions import DockerMachineError, Boot2DockerError
+from conductr_cli.exceptions import DockerMachineError, Boot2DockerError, MalformedBundleError
 from subprocess import CalledProcessError
 
 
@@ -104,6 +104,20 @@ def handle_bad_zip(func):
         try:
             return func(*args, **kwargs)
         except BadZipFile as err:
+            error('Problem with the bundle: {}', err.args[0])
+
+    # Do not change the wrapped function name,
+    # so argparse configuration can be tested.
+    handler.__name__ = func.__name__
+
+    return handler
+
+
+def handle_malformed_bundle(func):
+    def handler(*args, **kwargs):
+        try:
+            return func(*args, **kwargs)
+        except MalformedBundleError as err:
             error('Problem with the bundle: {}', err.args[0])
 
     # Do not change the wrapped function name,
